Updated data/shakespeare_char/prepare.py

+++ data/shakespeare_char/prepare.py
@@ -29,7 +29,7 @@ print(f"vocab size: {vocab_size:,}")
 stoi = { ch:i for i,ch in enumerate(chars) }
 itos = { i:ch for i,ch in enumerate(chars) }
 def encode(s):
-    return [stoi[c] for c in s] # encoder: take a string, output a list of integers
+    return [stoi[c] for c in s] # encoder: takes a string and outputs a list of integers
 def decode(l):
     ''.join([itos[i] for i in l]) # decoder: take a list of integers, output a string
 
@@ -45,8 +45,8 @@ print(f"train has {len(train_ids):,} tokens")
 print(f"val has {len(val_ids):,} tokens")
 
 # export to bin files
-train_ids = np.array(train_ids, dtype=np.int32)
-val_ids = np.array(val_ids, dtype=np.int32)
+train_ids = np.array(train_ids, dtype=np.uint16)
+val_ids = np.array(val_ids, dtype=np.uint16)
 train_ids.tofile(os.path.join(os.path.dirname(__file__), 'train.bin'))
 val_ids.tofile(os.path.join(os.path.dirname(__file__), 'val.bin'))
